Refactor claim list component into your claims page
