+++ src/js/disability-benefits/containers/YourClaimsPage.jsx
@@ -2,18 +2,78 @@ import React from 'react';
 import { connect } from 'react-redux';
 
 import Modal from '../../common/components/Modal';
-import { showConsolidatedMessage } from '../actions';
+import { getClaims, getFilteredClaims, changePage, showConsolidatedMessage } from '../actions';
 import AskVAQuestions from '../components/AskVAQuestions';
 import ConsolidatedClaims from '../components/ConsolidatedClaims';
 import FeaturesWarning from '../components/FeaturesWarning';
 import MainTabNav from '../components/MainTabNav';
+import ClaimsListItem from '../components/ClaimsListItem';
+import NoClaims from '../components/NoClaims';
+import Pagination from '../../common/components/Pagination';
+import LoadingIndicator from '../../common/components/LoadingIndicator';
+import { scrollToTop, setUpPage, setPageFocus } from '../utils/page';
 
 class YourClaimsPage extends React.Component {
+  constructor(props) {
+    super(props);
+    this.changePage = this.changePage.bind(this);
+  }
   componentDidMount() {
     document.title = 'Track Claims: Vets.gov';
+    this.loadClaims(this.props);
+    if (this.props.loading) {
+      scrollToTop();
+    } else {
+      setUpPage();
+    }
+  }
+  componentWillReceiveProps(newProps) {
+    if (this.props.allClaims && this.props.route.showClosedClaims !== newProps.route.showClosedClaims) {
+      this.loadClaims(newProps);
+      this.changePage(1);
+    }
+  }
+  componentDidUpdate(prevProps) {
+    if (!this.props.loading && prevProps.loading) {
+      setPageFocus();
+    }
+  }
+  loadClaims(props) {
+    if (props.allClaims) {
+      props.getFilteredClaims(!props.route.showClosedClaims);
+    } else {
+      props.getClaims();
+    }
+  }
+  changePage(page) {
+    this.props.changePage(page);
+    scrollToTop();
   }
 
   render() {
+    const { claims, pages, page, loading } = this.props;
+
+    let content;
+
+    if (loading) {
+      content = <LoadingIndicator message="Loading claims list" setFocus/>;
+    } else if (claims.length > 0) {
+      content = (<div className="claim-list">
+        {claims.map(claim => <ClaimsListItem claim={claim} key={claim.id}/>)}
+        <Pagination page={page} pages={pages} onPageSelect={this.changePage}/>
+      </div>);
+    } else {
+      content = <NoClaims/>;
+    }
+
+    if (this.props.allClaims) {
+      content = (
+        <div className="va-tab-content db-tab-content" role="tabpanel" id="tabPanelOpenClaims" aria-labelledby="tabOpenClaims">
+          {content}
+        </div>
+      );
+    }
+
     return (
       <div className="your-claims">
         <div className="row">
@@ -28,7 +88,7 @@ class YourClaimsPage extends React.Component {
               }}>Find out why we sometimes combine claims.</a>
             </p>
             {this.props.allClaims ? <MainTabNav/> : null}
-            {this.props.children}
+            {content}
             <Modal
                 onClose={() => true}
                 visible={this.props.consolidatedModal}
@@ -48,11 +108,18 @@ class YourClaimsPage extends React.Component {
 
 function mapStateToProps(state) {
   return {
+    loading: state.claims.list === null,
+    claims: state.claims.visibleRows,
+    pages: state.claims.pages,
+    page: state.claims.page,
     consolidatedModal: state.claims.consolidatedModal
   };
 }
 
 const mapDispatchToProps = {
+  getClaims,
+  getFilteredClaims,
+  changePage,
   showConsolidatedMessage
 };
